portfolio_risk_dashboard/requirements.txt
streamlit
yfinance
pandas
numpy
scipy
plotly
transformers
torch
requests
python-dotenv
